iot/requirements-dev.txt
google-genai
python-dotenv
google-generativeai
pyserial
Pillow
line-bot-sdk
requests
supabase
